[Game] updated: AI - added enum AiCommandCategory [Game] updated: UnitMovements - Flags are actually two flags Flags and actorFlags [Game] updated: TransferManager - added the ability to use the movement of vehicles from the path points parsed from the sniff-log [Game] updated: TimeManager - now the time on the server when it is turned on is synchronized with the current time [Game] updated: ZoneManager - local coordinates are now correctly converted into global coordinates. [Game] updated: Gimmick - now all the elevators are working properly - added the ability to stop in the middle [Data] updated: transfer_spawns.json [Login][G2L] rename LGRequestInfoPacket in GLRequestInfoPacket [Login][L2G] fixed: LGRequestInfoPacket

+++ AAEmu.Login/Core/Packets/L2G/LGRequestInfoPacket.cs
@@ -7,7 +7,7 @@ namespace AAEmu.Login.Core.Packets.L2G
     {
         private readonly uint _connectionId;
         private readonly uint _requestId;
-        private readonly ulong _accountId;
+        private readonly uint _accountId;
 
         public LGRequestInfoPacket(uint connectionId, uint requestId, uint accountId) : base(0x03)
         {
